Use symlinks for sun SUNWCCh files
The SUNWCCh files are additional files needed to intercept the
search for system headers on Sun, otherwise the platform-supplied
SUNWCCh files are found by searching the entire path before the
compiler falls back on actually searching for the header requested
by its exact name.

As the SUNWCCh files are supposed to be maintained in synch with
our intercept headers, we prefer to guarantee consistency by
using symlinks.  An experiment in this direction was tried once
before a few years ago, but was backed out without recording why
it failed.  Testing again with the current Sun Studion compiler
seems to supprt symlinks, so we will try again.