Testing Sun symlinks without #include_next
This branch has effectively two patches:
1. Switch .SUNWCCh files over to symlinks, to pick up any
bugfixes already applied to regular headers, but not yet
copied back for Sun.
2. Adopt #include_next support in the SunCC compiler to
address some platform issues in later compilers.

It is noted that the #include_next support does not seem
to work as expected, and fails to build (but not because
the compiler cannot parse #include_next) so this first
patch confirms that the symlinks work, before an update
that tries to properly support #include_next on this
platform.