Cleanup after #include_next fails on Sun
The #include_next feature on Sun does not work in a way that our existing
code that uses the feature expects.  This patch cleans up back to a
stable state after removing that support, but retaining ancilliary fixes
found along the way.

For reference, #include_next on Sun CC will simply search for the named
file, whether or not it has a name that matches the current file,
resuming its search from the search path that found the current file,
even if it is a .cpp (why test driver fails).  I have not yet confirmed
why the bsl+stdhdrs fail to include_next, but it seems to be related to
the same issue.