Fixed for bitwise movable trait
Following many of the same issues reported for the is trivially
copyable trait.  This patch addresses array of bitwise moveable
types, and cv-qualified bitwise movable types.  Note that many
of the fixes for trivially copyable types are inherited through
that metafunction, but we still need to fix the case of users
specializing this trait directly.

In addition, this component was essentially missing an entire
test driver, so provide the missing test cases inspired by
the similar work on trivially copyable.

Finally, we note that the usage example is present ONLY in the
test driver, and has not been copied into the header.  We
deliberately defer fixing that, as reviewing and moving the
usage example should be a separate issue.