Fixed for is_trivially_default_constructible
Following the earlier examples of trivially copyable and
bitwise movebale traits, this patch corrects the is
trivially default constructible trait to correctly handle
arrays and cv-qualified types, and especially explicit
user specialization of the trait in such cases.

The test driver is extended to properly test the range
of type concerns for this trait.