Apply workaround for defective compilers
After committing the is_trivial patches to dev-integration,
the nightly tests failed for the usual suspects of Sun and
IBM.  Investigating further revealed a number of components
in 'bslmf' that have less than stellar test drivers, so
several additional components have had their test drivers
updated in order to find the necessary fixes.

One thing this patch does NOT fix is that BDE and the C++
standard disagree on whether 'is_convertible' should allow
any type to convert to cv-'void'.  The C++ standard says
that all types convert to 'void', where BDE requires that
only cv-'void' types convert to 'void'.  This should be
the subject of a follow-up ticket, as it is potentially
too risky a change to make as a drive-by fix on this
branch.