Additional fix for is_convertible from void
Merge conflicts trying to reach dev-integration showed another patch in flight
addressing concerns with converting from 'void' types, that showed an
ambiguity with converting cv-'void' to a reference type, due to other
matches that partially specialize on converting to references.

It also revealed some missing tests for function and function reference
types that should have been part of the original set in this branch as
well.