Revised workarounds for traits on IBM/Sun
Trying to support matching function types is still causing the Sun CC
and IBM xlC compilers to fail to compile, complaining about our syntax.
Fall back to plan B, and simply hack around the problem of dispatching
function types into the matcher in the first place.  It is likely that
when we have more extensive testing through the whole of 'bslmf', this
hack will be needed in a couple more places.